% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_density}
\alias{calc_density}
\title{Calculate the densities of each metabolite in sdata}
\usage{
calc_density(sdata)
}
\arguments{
\item{sdata}{An sdata list containing WT and MUT dataframes}
}
\value{
A means list containing WT and MUT mean dataframes
}
\description{
This function calculates the density for each column (=exchange reaction = metabolite).
}
