% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate the difference between random pairs of flux samples}
\usage{
calc_diff(sdata, seed = NA)
}
\arguments{
\item{sdata}{An sdata list containing WT and MUT dataframes}

\item{seed}{A seed to set for reproducibility}
}
\value{
A "long" dataframe containing all diff samples for all metabolites
}
\description{
This function uses sdata to calculate random subtractions between c2 and c1 sampling results.
}
