% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_filter_density}
\alias{calc_filter_density}
\title{Filter the metabolites based on their zscore}
\usage{
calc_filter_density(sdata, zscores, thr = 2, max_n = 10)
}
\arguments{
\item{sdata}{An sdata list containing WT and MUT dataframes}
}
\value{
A means list containing WT and MUT mean dataframes
}
\description{
This function filters metabolites based on their previously calculated zscore using densities.
}
