% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_sdata_means}
\alias{calc_sdata_means}
\title{Calculate the mean of each exchange reaction in sdata}
\usage{
calc_sdata_means(sdata)
}
\arguments{
\item{sdata}{An sdata list containing WT and MUT dataframes}
}
\value{
A means list containing WT and MUT mean dataframes
}
\description{
This function calculates the mean for each column (=exchange reaction = metabolite) and transforms the data into a
format usable by plot_scatter.
}
