% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamba.R
\name{getSamba}
\alias{getSamba}
\title{Title}
\usage{
getSamba(version = NA, bfc = NULL)
}
\arguments{
\item{bfc}{}
}
\value{
A message with the version to retrieve
}
\description{
Title
}
\examples{
getSamba("v0.1.0")
}
