% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_shiny_sampling_results}
\alias{load_shiny_sampling_results}
\title{Read in two conditions of sampling data within a Shiny app}
\usage{
load_shiny_sampling_results(filepath, prefix)
}
\arguments{
\item{prefix}{Prefix of the two files you are inputting: should be the same for both}

\item{indir}{Path to the input dir, should contain a ko folder and wt folder if not using single.wt}
}
\value{
A list containing WT and MUT dataframes of all samples
}
\description{
This function reads in two conditions of sampling data. It assumes that the two results files have the same prefix
(disease) and two different suffixes corresponding to the conditions, one of which must be "WT".
}
