% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_basic_distrib}
\alias{plot_basic_distrib}
\title{Plot a basic demo distribution}
\usage{
plot_basic_distrib()
}
\description{
Compare WT and MUT flux distributions for the top set of zscore metabolites for multiple diseases.
}
