% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_demo_distrib}
\alias{plot_demo_distrib}
\title{Plot a basic demo distribution}
\usage{
plot_demo_distrib()
}
\description{
Compare WT and MUT flux distributions for the top set of zscore metabolites for an example disease.
}
