% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distrib}
\alias{plot_distrib}
\title{Plot distributions for 1 case from raw samples}
\usage{
plot_distrib(sdata, zscore, thr, max_n = 10, case = "Disease")
}
\arguments{
\item{sdata}{A list containing a WT dataframe and a MUT dataframe. Each dataframe contains X samples for metabolites}

\item{zscore}{A dataframe of zscore calculated using calc_zscore}

\item{thr}{The threshold for filtering zscores}

\item{max_n}{The max number of metabolites to show}

\item{case}{The name of the disease or condition}
}
\description{
Compare WT and MUT flux distributions for the top set of zscore metabolites.
}
