% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_zscore_distrib}
\alias{plot_zscore_distrib}
\title{Plot a histogram of z-scores}
\usage{
plot_zscore_distrib(
  zscore,
  metab_dict,
  thr,
  case = "Disease",
  binwidth = 0.1,
  labels = T
)
}
\arguments{
\item{zscore}{A dataframe of zscore calculated using calc_zscore}

\item{metab_dict}{A dataframe containing an ID column and a name column. This is useful for converting metabolite IDs
to readable names. It can be generated using sambapy from the metabolic model.}

\item{thr}{The threshold for filtering zscores}

\item{case}{The name of the disease or condition}
}
\description{
Plots a histogram of z-scores and highlights the z-scores above the threshold.
}
