% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_lyrics.R
\name{translate_lyrics}
\alias{translate_lyrics}
\title{Translate song lyrics}
\usage{
translate_lyrics(artist, song, target = "en", destdir = ".", ...)
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization
does not. See \code{genius::genius_lyrics()}}

\item{song}{The quoted name of the song. Spelling matters, capitalization
does not. See \code{genius::genius_lyrics()}}

\item{target}{The target language}

\item{...}{Other arguments for gl_translate}
}
\value{
A two-column tibble with the original and the translated lyrics
}
\description{
This function was written out of saudade to
translate any lyrics available in the Genius.com platform. It will save the
lyrics and the translation to avoid repeated queries to both databases
}
\examples{
\dontrun{
translate_lyrics(artist = "caetano veloso", song = "é hoje", target = "en")
translate_lyrics(artist = "caetano veloso", song = "é hoje", target = "es")
}
}
