% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate the difference between random pairs of flux samples}
\usage{
calc_diff(sdata, seed = NA, divfactor = 1)
}
\arguments{
\item{sdata}{An sdata list containing c1 and c2 dataframes}

\item{seed}{A seed to set for reproducibility}

\item{divfactor}{A division factor to divide the number of samples by for less intensive calculations}
}
\value{
A "long" dataframe containing all diff samples for all metabolites
}
\description{
This function uses sdata to calculate random subtractions between c2 and c1 sampling results.
}
