% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_multi_density_distrib}
\alias{plot_multi_density_distrib}
\title{Plot multiple density distributions}
\usage{
plot_multi_density_distrib(densities_filt, metab_dict, thr = 2, max_n = 10)
}
\arguments{
\item{densities_filt}{A list containing a WT list and a MUT list. Each list contains X densities for metabolites}

\item{zscore}{A dataframe of zscore calculated using calc_zscore}
}
\description{
Compare WT and MUT flux distributions for the top set of zscore metabolites for multiple diseases using filtered
densities as inputs instead of raw sdata.
}
