% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot a scatter plot of WT means vs MUT means}
\usage{
plot_scatter(
  sdata,
  zscore,
  means,
  metab_dict,
  thr,
  case = "Disease",
  outlier.dist = 5,
  outlier.count = 0
)
}
\arguments{
\item{sdata}{A list containing a WT dataframe and a MUT dataframe. Each dataframe contains X samples for metabolites}

\item{zscore}{A dataframe of zscore calculated using calc_zscore}

\item{thr}{The threshold for filtering zscores}

\item{case}{The name of the disease or condition}
}
\description{
Plots a scatter plot with WT flux means on the x-axis and MUT flux means on the y-axis. The points are coloured
by zscore.
}
